// src/components/Forms/Form.types.ts
export type HeadingType = 'h2' | 'h3' | 'h4';
export enum FormBgColor {
  Grey = 'bg-grey-light',
  White = 'bg-white',
  Navy = 'bg-bright-navy',
  Green = 'bg-challenger-green',
  Blue = 'bg-blue',
  Teal = 'bg-deep-teal',
  DeepBlue = 'bg-deep-blue',
  LightBlue = 'bg-light-blue',
  None = '',
}